Separate Bookmarksettings, and update docs
