+++ docs/settingup.md
@@ -311,8 +311,8 @@ Apply a bookmark in the current app.
 
 ### Settings
 
-* `title`: (optional) Name of the bookmark to apply.
-* `id`: (optional) GUID of the bookmark to apply.
+* `title`: Name of the bookmark.
+* `id`: GUID of the bookmark.
 
 ### Example
 
@@ -396,9 +396,9 @@ Create a bookmark from the current selection and selected sheet.
 
 ### Settings
 
-* `title`: Name of the bookmark to create.
+* `title`: Name of the bookmark.
+* `id`: GUID of the bookmark.
 * `description`: (optional) Description of the bookmark to create.
-* `id`: (optional) ID to use with subsequent `applybookmark` or `deletebookmark` actions. **Note:** This ID is only used within the scenario.
 * `nosheet`: Do not include the sheet location in the bookmark.
 * `savelayout`: Include the layout in the bookmark.
 
@@ -449,12 +449,12 @@ Delete one or more bookmarks in the current app.
 
 ### Settings
 
+* `title`: Name of the bookmark.
+* `id`: GUID of the bookmark.
 * `mode`: 
     * `single`: Delete one bookmark that matches the specified `title` or `id` in the current app.
     * `matching`: Delete all bookmarks with the specified `title` in the current app.
     * `all`: Delete all bookmarks in the current app.
-* `title`: (optional) Name of the bookmark to delete.
-* `id`: (optional) GUID of the bookmark to delete.
 
 ### Example
 
@@ -1797,6 +1797,15 @@ Upload a data file to the Data manager.
 ```
 
 </details>
+</details>
+<details>
+<summary>Ungrouped actions</summary>
+
+
+* `title`: Name of the bookmark.
+* `id`: GUID of the bookmark.
+
+
 </details>
 
 ## Session variables
