+++ generatedocs/data/params.json
@@ -213,23 +213,17 @@
     "config.settings.outputs.dir": [
         "Directory in which to save artifacts generated by the script (except log file)."
     ],
-    "applybookmark.title": [
-        "(optional) Name of the bookmark to apply."
+    "bookmark.title": [
+        "Name of the bookmark."
     ],
-    "applybookmark.id": [
-        "(optional) GUID of the bookmark to apply."
+    "bookmark.id": [
+        "GUID of the bookmark."
     ],
     "changesheet.id": [
         "GUID of the sheet to change to."
     ],
-    "createbookmark.title": [
-        "Name of the bookmark to create."
-    ],
     "createbookmark.description": [
         "(optional) Description of the bookmark to create."
-    ],
-    "createbookmark.id": [
-        "(optional) ID to use with subsequent `applybookmark` or `deletebookmark` actions. **Note:** This ID is only used within the scenario."
     ],
     "createbookmark.nosheet": [
         "Do not include the sheet location in the bookmark."
@@ -251,12 +245,6 @@
         "`single`: Delete one bookmark that matches the specified `title` or `id` in the current app.",
         "`matching`: Delete all bookmarks with the specified `title` in the current app.",
         "`all`: Delete all bookmarks in the current app."
-    ],
-    "deletebookmark.title": [
-        "(optional) Name of the bookmark to delete."
-    ],
-    "deletebookmark.id": [
-        "(optional) GUID of the bookmark to delete."
     ],
     "deletedata.filename": [
         "Name of the file to delete."
